Refactor EventCategory logic into data/event;
